fix(api): convert anonymous funcs to M:F
Anonymous functions would crashes with error `badfun` if the code has
been changed. Especially the anonymous functions is sent to processes
in a message, passed as an argument of an RPC call, or stored
in ETS tables.

For anonymous funcs in RPC calls and messages, we could change it to
equivalent M:F:A calls.

For those stored in ETS tables, we could rebuild it when the exception
arises.